requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
Pillow==8.3.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.7
